OK this is a lot of stuff but very little implementation. Hopefully, I'll get it all here but just in case look over all of the updated files for their changes in this commit and the previous one or two. Sorry, Next time I'll do a big change like this that needs the documentation in one sweep.
IController interface is done as I see it

GuiView - THis is the new refractored GUIView.

GuiViewFrame - This has the "implemented" methods sans implementation

ViewFactory - This has been modified to contain the CompositeView

MidiViewImpl - This has been modified so the run only plays one beats notes. As I understand it. The controller will recall run on the composite view which will call run on the midi, and guiView every tick.
I'm still not sure if that is actually the course of action that we want to take.